SMS/requirements.txt
pillow
django-crispy-forms
foundation
six